requirements.txt
Flask-Cors
Flask-SQLAlchemy
flask_jwt_extended
python-decouple
Flask-Migrate
Werkzeug